requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
huggingface_hub
numpy
pypdf
pymupdf
pydantic
tenacity
chromadb
langchain-chroma
langchain-community
langchain-core
langchain-text-splitters
rank_bm25
sentence-transformers
streamlit
